update 06-register 24.08 16:52

+++ 06-register/note.controller.js
@@ -1,26 +1,34 @@
-const chalk = require('chalk')
-const Note = require('./models/Note')
+const chalk = require('chalk');
+const Note = require('./models/Note');
 
-const removeNote = async (id) => {
-  await Note.deleteOne({ _id: id })
-  console.log(chalk.green('OK. Note was removed'))
-}
+const removeNote = async (id, owner) => {
+  const result = await Note.deleteOne({ _id: id, owner });
 
-const updateNote = async (id, title) => {
-  await Note.updateOne({ _id: id }, { title })
-  console.log(chalk.green('OK. Note was updated'))
-}
+  if (result.deletedCount === 0) {
+    throw new Error('You cannot delete this note');
+  }
+  console.log(chalk.green('OK. Note was removed'));
+};
+
+const updateNote = async (id, title, owner) => {
+  const result = await Note.updateOne({ _id: id, owner }, { title });
+
+  if (result.matchedCount === 0) {
+    throw new Error('You cannot update this note');
+  }
+  console.log(chalk.green('OK. Note was updated'));
+};
 
 const addNote = async (title, owner) => {
-  await Note.create({ title, owner })
-  console.log(chalk.green('OK. Note was added'))
-}
+  await Note.create({ title, owner });
+  console.log(chalk.green('OK. Note was added'));
+};
 
-const getNotes = async () => await Note.find()
+const getNotes = async () => await Note.find();
 
 module.exports = {
   addNote,
   removeNote,
   getNotes,
   updateNote,
-}
+};
